Add clippy enforcement to AoC workflow + fix all warnings
- Fix day08.rs: Use iterator instead of range loop
- Fix day09.rs: Remove redundant trim() before split_whitespace()
- Fix day09.rs: Use array instead of vec! for fixed-size data
- Fix day10.rs: Remove unnecessary cast (cycle % 40)
- Update aoc-instructions.md: Add clippy as step 6 in workflow
- Update aoc-instructions.md: Add clippy to verification checklist
- All clippy warnings now fixed: cargo clippy -p aoc2022 --all-targets -- -D warnings passes

+++ advent_of_code/aoc2022/src/solver/day10.rs
@@ -83,7 +83,7 @@ fn solve_part2_with_instructions(instructions: &[Instruction]) -> String {
         
         for _ in 0..cycles_to_execute {
             // Calculate CRT column position (0-39 within current row)
-            let crt_col = (cycle % 40) as i32;
+            let crt_col = cycle % 40;
             
             // Sprite is 3 pixels wide, centered on X: [X-1, X, X+1]
             // Draw '#' if CRT position overlaps with sprite
